backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.41.2",
    "fastapi>=0.121.3",
    "ffmpeg-python>=0.2.0",
    "llama-index>=0.14.8",
    "llama-index-llms-anthropic>=0.10.2",
    "openai-whisper>=20250625",
    "pymysql>=1.1.2",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "yt-dlp>=2025.11.12",
]


